Move batching to Linux TUN impl
