+++ rust/libs/connlib/tunnel/src/device_channel.rs
@@ -3,9 +3,6 @@ use std::io;
 use std::task::{Context, Poll, Waker};
 use tun::Tun;
 
-#[cfg(target_os = "linux")]
-use tun::IpPacketBatch;
-
 pub struct Device {
     tun: Option<Box<dyn Tun>>,
     waker: Option<Waker>,
@@ -62,12 +59,6 @@ impl Device {
 
         self.tun()?.send(packet)?;
 
-        Ok(())
-    }
-
-    #[cfg(target_os = "linux")]
-    pub fn send_batch(&mut self, batch: IpPacketBatch) -> io::Result<()> {
-        self.tun()?.send_batch(batch)?;
         Ok(())
     }
